bot/requirements.txt
python-telegram-bot==13.15
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0



